migrating voting classifier to classifiers package
