+++ eta/classifiers/__init__.py
@@ -8,4 +8,6 @@ Brian Moore, brian@voxel51.com
 '''
 
 # Import all classifiers into the `eta.classifiers` namespace
-from vgg16 import VGG16Classifier, VGG16ClassifierConfig
+from vgg16_classifiers import VGG16Classifier, VGG16ClassifierConfig
+from voting_classifiers import VideoFramesVotingClassifier, \
+                               VideoFramesVotingClassifierConfig
