+++ eta/core/learning.py
@@ -20,7 +20,6 @@ from builtins import *
 # pragma pylint: enable=wildcard-import
 
 from eta.core.config import Config, ConfigError, Configurable
-import eta.core.data as etad
 import eta.core.utils as etau
 
 
@@ -59,9 +58,6 @@ def write_labels_map(labels_map, outpath):
     with open(outpath, "w") as f:
         for idx in sorted(labels_map):
             f.write("%s:%s\n" % (idx, labels_map[idx]))
-
-
-# MODEL INTERFACES ############################################################
 
 
 class ModelConfig(Config):
@@ -557,60 +553,3 @@ class VideoObjectDetector(Detector):
                 the detections for the video
         '''
         raise NotImplementedError("subclasses must implement detect()")
-
-
-# CONCRETE CLASSIFIERS ########################################################
-
-
-class VideoFramesVotingClassifierConfig(Config):
-    '''Configuration settings for a VideoFramesVotingClassifier.'''
-
-    def __init__(self, d):
-        self.image_classifier = self.parse_object(
-            d, "image_classifier", ImageClassifierConfig)
-        self.confidence_weighted_vote = self.parse_bool(
-            d, "confidence_weighted_vote", default=False)
-
-
-class VideoFramesVotingClassifier(VideoFramesClassifier):
-    '''A video frames classifier that uses an `ImageClassifier` to classify
-    each image and then votes on each attribute to determine the predictions
-    for the video.
-
-    Note that all attributes are combined into a single vote. Thus, even if the
-    `ImageClassifier` is a multilabel classifier, each prediction will contain
-    the single most prevelant label.
-    '''
-
-    def __init__(self, config):
-        '''Creates a VideoFramesVotingClassifier instance.
-
-        Args:
-            config: a VideoFramesVotingClassifierConfig instance
-        '''
-        self.config = config
-        self.image_classifier = config.image_classifier.build()
-
-    def __enter__(self):
-        self.image_classifier.__enter__()
-        return self
-
-    def __exit__(self, *args):
-        self.image_classifier.__exit__(*args)
-
-    def predict(self, imgs):
-        '''Peforms prediction on the given video represented as a tensor of
-        images.
-
-        Args:
-            imgs: a list (or d x ny x nx x 3 tensor) of images defining the
-                video to classify
-
-        Returns:
-            an `eta.core.data.AttributeContainer` instance describing
-                the predictions for the input
-        '''
-        frame_attrs = self.image_classifier.predict_all(imgs)
-        return etad.majority_vote_categorical_attrs(
-            frame_attrs,
-            confidence_weighted=self.config.confidence_weighted_vote)
